Add ScheduleSQL to MetricStore parameters
Link: https://code.aone.alibaba-inc.com/sls/aliyun-log-java-sdk/codereview/5598629

* add metric parameter

* change fromTime&toTime int to long

* fix json key bug

* optimize ScheduledSQLConfiguration

* fix deserialize bug

* from/toTime legitimacy judgment

* append method

* add no para construction

* add builder

* optimize code

* return void

* optimize builder

* fix parameter key

+++ src/main/java/com/aliyun/openservices/log/Client.java
@@ -5010,10 +5010,10 @@ public class Client implements LogService {
 	@Override
 	public CreateScheduledSQLResponse createScheduledSQL(CreateScheduledSQLRequest request) throws LogException{
 		ScheduledSQL scheduledSQL = (ScheduledSQL) request.getBody();
-		Integer fromTime = scheduledSQL.getConfiguration().getFromTime();
-        Integer toTime = scheduledSQL.getConfiguration().getToTime();
-		boolean timeRange = fromTime > 1451577600 && toTime > fromTime ;
-		boolean sustained = fromTime > 1451577600 && toTime == 0;
+		Long fromTime = scheduledSQL.getConfiguration().getFromTime();
+        Long toTime = scheduledSQL.getConfiguration().getToTime();
+		boolean timeRange = fromTime > 1451577600L && toTime > fromTime;
+		boolean sustained = fromTime > 1451577600L && toTime == 0;
 		if ((!timeRange) && (!sustained)){
 			throw new IllegalArgumentException("Invalid fromTime: "+fromTime+" toTime: "+toTime+
                     ", please ensure fromTime more than 1451577600.");
